requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
PyPDF2==3.0.1
numpy==2.1.3
tiktoken==0.6.0 